# app/static/swagger.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: "Cinema API"
  description: "This is a Cinema API where you can execute CRUD operations on 
  movies, actors, genres, sessions, customers and auditoriums"
servers:
  - url: http://127.0.0.1:5000
...
paths:
...
  /users:
    get:
      tags:
        - "Users"
      security:
        - bearerAuth: [ ]
      summary: "Get users information"
      description: "This can only be done by the logged in admin"
      parameters:
        - name: "firstname"
          in: "query"
          description: "Firstname of user"
          required: false
          schema:
            type: "string"
        - name: "lastname"
          in: "query"
          description: "Lastname of user"
          required: false
          schema:
            type: "string"
        - name: "email"
          in: "query"
          description: "Email of user"
          required: false
          schema:
            type: "string"
        - name: "offset"
          in: "query"
          description: "Offset"
          required: false
          schema:
            type: "string"
        - name: "limit"
          in: "query"
          description: "Limit"
          required: false
          schema:
            type: "string"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersOut'
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
    post:
      security:
        - bearerAuth: []
      tags:
        - "Users"
      summary: "Create new user"
      description: "This can only be done by the logged in admin"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserIn'
      responses:
        '201':
          description: "Successful Operation"
          content:
            application/json:
              example:
                id: "5"
        '400':
          description: "Wrong input data"
          content:
            application/json:
              example:
                message: "Please, specify 'firstname', 'lastname', 'email', 'password' and 'is_admin'"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "Wrong input email"
          content:
            application/json:
              example:
                message: "Email {email} already used"
  /users/{id}:
    get:
      tags:
        - "Users"
      security:
        - bearerAuth: []
      parameters:
        - name: "id"
          in: "path"
          description: "User ID"
          required: true
          schema:
            type: "integer"
      summary: "Get user by id"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "User not found"
          content:
            application/json:
              example:
                message: "User not found"
    patch:
      security:
        - bearerAuth: []
      tags:
        - "Users"
      summary: "Update user's information"
      description: "This can only be done by the logged in user"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of actor"
          required: true
          schema:
            type: "integer"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserIn'
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              example:
                message: "Updated"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "Wrong user's id"
          content:
            application/json:
              example:
                message: "User not found"
    delete:
      security:
        - bearerAuth: []
      tags:
        - "Users"
      summary: "Delete user"
      description: "This can only be done by the logged in user"
      parameters:
        - name: "id"
          in: "path"
          description: "Id of the user"
          required: true
          schema:
            type: "integer"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              example:
                message: "Deleted"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "Wrong user's id"
          content:
            application/json:
              example:
                message: "User not found"
        '405':
          description: "Admin access required"
          content:
            application/json:
              example:
                message: "Not allowed"
  /users/inactive:
    get:
      tags:
        - "Users"
      security:
        - bearerAuth: []
      summary: "Get inactive users information"
      description: "This can only be done by the logged in admin"
      parameters:
        - name: "offset"
          in: "query"
          description: "Offset"
          required: false
          schema:
            type: "string"
        - name: "limit"
          in: "query"
          description: "Limit"
          required: false
          schema:
            type: "string"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersOut'
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserIn:
      type: "object"
      properties:
        firstname:
          type: "string"
          example: "John"
        lastname:
          type: "string"
          example: "Doe"
        email:
          type: "string"
          example: "test1@gmail.com"
        password:
          type: "string"
          example: "password"
        is_admin:
          type: "boolean"
          example: false
...
    UserOut:
      type: "object"
      properties:
        id:
          type: "integer"
          example: 2
        firstname:
          type: "string"
          example: "John"
        lastname:
          type: "string"
          example: "Doe"
        email:
          type: "string"
          example: "sample@mail.com"
        is_admin:
          type: "boolean"
          example: false
        is_active:
          type: "boolean"
          example: true
    UsersOut:
      type: "array"
      items:
        $ref: '#/components/schemas/UserOut'